refactor(utils): remove stream utils to internals
